Clean up file filters
